+++ web/src/main/ts/Sidebar.tsx
@@ -5,10 +5,10 @@ import { SelectToolSidebar } from "./SelectToolSidebar";
 import { ContourList } from "./ContourList";
 
 interface SidebarProps {
-    tool: string,
-    contours: any,
-    selectContour: (index:number) => void,
-    selectedContour: number,
+    tool: string;
+    contours: any;
+    selectContour: (index:number) => void;
+    selectedContour: number;
 }
 
 export class Sidebar extends React.Component<SidebarProps,{}> {